move useTokenBalance to wallet
